ajs_article {rticles}	R Documentation

Austrian Journal of Statistics (AJS) format.

Description

Format for creating a Austrian Journal of Statistics (AJS) article. Adapted from https://www.jstatsoft.org/about/submissions.

Usage

ajs_article(
  ...,
  keep_tex = TRUE,
  citation_package = "natbib",
  pandoc_args = NULL
)

Arguments

...
Arguments to rmarkdown::pdf_document()

keep_tex
Keep the intermediate tex file used in the conversion to PDF

citation_package
The LaTeX package to process citations, natbib or biblatex. Use default if neither package is to be used, which means citations will be processed via the command pandoc-citeproc.

pandoc_args
Additional command line options to pass to pandoc

[Package rticles version 0.21 Index]