amq_article {rticles}	R Documentation

Format pour Bulletin de l'AMQ.

Description

Ce format a été adapté du format du bulletin de l'AMQ

Usage

amq_article(..., latex_engine = "xelatex", keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris"), fig_caption = TRUE)

Arguments

...
Arguments to rmarkdown::pdf_document

latex_engine
LaTeX engine for producing PDF output. Options are "pdflatex", "lualatex", and "xelatex".

keep_tex
Keep the intermediate tex file used in the conversion to PDF

md_extensions
Markdown extensions to be added or removed from the default definition or R Markdown. See the rmarkdown_format for additional details.

fig_caption
TRUE to render figures with captions

Value

R Markdown output format to pass to render

Examples

## Not run: 
rmarkdown::draft("MyArticle.Rmd", template = "amq_article", package = "rticles")

## End(Not run)

[Package rticles version 0.6 Index]