ams_article {rticles}	R Documentation

American Meteorological Society (AMS) Journal format.

Description

Format for creating an American Meteorological Society (AMS) Journal articles. Adapted from https://www.ametsoc.org/ams/index.cfm/publications/authors/journal-and-bams-authors/author-resources/latex-author-info/.

Usage

ams_article(..., keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris"))

Arguments

...
Arguments to rmarkdown::pdf_document

keep_tex
Keep the intermediate tex file used in the conversion to PDF

md_extensions
Markdown extensions to be added or removed from the default definition or R Markdown. See the rmarkdown_format for additional details.

Value

R Markdown output format to pass to render

Examples

## Not run: 
library(rmarkdown)
draft("MyArticle.Rmd", template = "ams_article", package = "rticles")

## End(Not run)

[Package rticles version 0.6 Index]