biometrics_article {rticles}	R Documentation

Biometrics Journal format.

Description

This format was adapted from the Biometrics Macro package.

Usage

biometrics_article(..., keep_tex = TRUE, citation_package = "natbib")

Arguments

...
Arguments to rmarkdown::pdf_document

keep_tex
Keep the intermediate tex file used in the conversion to PDF

citation_package
The LaTeX package to process citations, natbib or biblatex. Use none if neither package is to be used.

Value

R Markdown output format to pass to render

Examples

## Not run: 
rmarkdown::draft("MyArticle.Rmd", template = "biometrics_article", package = "rticles")

## End(Not run)

[Package rticles version 0.6 Index]