Copernicus journals format.

Format for creating submissions to Copernicus journals.

copernicus_article(
  ...,
  keep_tex = TRUE,
  highlight = NULL,
  citation_package = "natbib",
  md_extensions = c("-autolink_bare_uris", "-auto_identifiers")
)

copernicus_journal_abbreviations(journal_name = "*")

Arguments

...
Additional arguments to rmarkdown::pdf_document(). Note: extra_dependencies are not allowed as Copernicus does not support additional packages included via \usepackage{}.

keep_tex
Keep the intermediate tex file used in the conversion to PDF

highlight
Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock". Pass NULL to prevent syntax highlighting.

citation_package
The LaTeX package to process citations, natbib or biblatex. Use default if neither package is to be used, which means citations will be processed via the command pandoc-citeproc.

md_extensions
Markdown extensions to be added or removed from the default definition or R Markdown. See the rmarkdown_format for additional details.

journal_name
A regular expression to filter the by the journal name, see pattern in grep; defaults to *.

Value

An R Markdown output format.

Details

This was adapted from https://publications.copernicus.org/for_authors/manuscript_preparation.html.

An number of required and optional manuscript sections, e.g. acknowledgements, competinginterests, or authorcontribution, must be declared using the respective properties of the R Markdown header - see skeleton file.

Version: Based on copernicus_package.zip in the version 6.2, 15 January 2021, using copernicus.cls in version 9.25.

Copernicus journal abbreviations: You can use the function copernicus_journal_abbreviations() to get the journal abbreviation for all journals supported by the Copernicus article template.

Important note: The online guidelines by Copernicus are the official resource. Copernicus is not responsible for the community contributions made to support the template in this package. Copernicus converts all typeset TeX files into XML, the expressions and markups have to be highly standardized. Therefore, please keep the following in mind:

Please provide only one figure file for figures with several panels, and please do not use \subfloat or similar commands.

Please use only commands in which words, numbers, etc. are within braces (e.g. \textrm{TEXT} instead of {\rm TEXT}).

For algorithms, please use the syntax given in template.tex or provide your algorithm as a figure.

Please do not define new commands.

Supported packages (\usepackage{}) are already integrated in the copernicus.cls. Please do not insert additional ones in your *.tex file.

If you opt for syntax highlighting for your preprint or other reasons, please do not forget to use highlight = NULL for your final file upload once your manuscript was accepted for publication.

Spaces in labels (\label{}) are not allowed; please make sure that no label name is assigned more than once.

Please do not use \paragraph{}; only \subsubsection{} is allowed.

It is not possible to add tables in colour.

Note

If you use rmarkdown::pdf_document(), all internal references (i.e. tables and figures) must use \ref{} whereas with bookdown::pdf_document2(), you can additionally use \@ref().

References

Manuscript preparation guidelines for authors. https://publications.copernicus.org/for_authors/manuscript_preparation.html

Examples

names(copernicus_journal_abbreviations())
#>  [1] "Advances in Geosciences"                                              
#>  [2] "Advances in Radio Science"                                            
#>  [3] "Advances in Science and Research"                                     
#>  [4] "Advances in Statistical Climatology, Meteorology and Oceanography"    
#>  [5] "Annales Geophysicae"                                                  
#>  [6] "Archives Animal Breeding"                                             
#>  [7] "ASTRA Proceedings"                                                    
#>  [8] "Atmospheric Chemistry and Physics"                                    
#>  [9] "Atmospheric Measurement Techniques"                                   
#> [10] "Biogeosciences"                                                       
#> [11] "Climate of the Past"                                                  
#> [12] "DEUQUA Special Publications"                                          
#> [13] "Drinking Water Engineering and Science"                               
#> [14] "European Journal of Mineralogy"                                       
#> [15] "Earth Surface Dynamics"                                               
#> [16] "Earth System Dynamics"                                                
#> [17] "Earth System Science Data"                                            
#> [18] "E&G Quaternary Science Journal"                                       
#> [19] "Fossil Record"                                                        
#> [20] "Geochronology"                                                        
#> [21] "Geographica Helvetica"                                                
#> [22] "Geoscience Communication"                                             
#> [23] "Geoscientific Instrumentation, Methods and Data Systems"              
#> [24] "Geoscientific Model Development"                                      
#> [25] "History of Geo- and Space Sciences"                                   
#> [26] "Hydrology and Earth System Sciences"                                  
#> [27] "Journal of Bone and Joint Infection"                                  
#> [28] "Journal of Micropalaeontology"                                        
#> [29] "Journal of Sensors and Sensor Systems"                                
#> [30] "Magnetic Resonance"                                                   
#> [31] "Mechanical Sciences"                                                  
#> [32] "Natural Hazards and Earth System Sciences"                            
#> [33] "Nonlinear Processes in Geophysics"                                    
#> [34] "Ocean Science"                                                        
#> [35] "Primate Biology"                                                      
#> [36] "Proceedings of the International Association of Hydrological Sciences"
#> [37] "Scientific Drilling"                                                  
#> [38] "SOIL"                                                                 
#> [39] "Solid Earth"                                                          
#> [40] "The Cryosphere"                                                       
#> [41] "Web Ecology"                                                          
#> [42] "Weather and Climate Dynamics"                                         
#> [43] "Wind Energy Science"                                                  
copernicus_journal_abbreviations(journal_name = "Science Data")
#> Earth System Science Data 
#>                    "essd" 
if (FALSE) {
library("rmarkdown")
draft("MyArticle.Rmd", template = "copernicus", package = "rticles")
render("MyArticle/MyArticle.Rmd")
}