ctex {rticles}	R Documentation

A PDF format for documents based on the LaTeX package ctex

Description

ctex() is a wrapper function for rmarkdown::pdf_document() and changed the default values of two arguments template and latex_engine so it works better with the ctex package.

Usage

ctex(..., template = ctex_template(), latex_engine = "xelatex")

ctex_template()

Arguments

..., template, latex_engine
Passed to markdown::pdf_document()

Value

ctex() returns a format that can be passed to rmarkdown::render(); ctex_template() returns the path to a LaTeX template in rticles for Chinese documents using the ctex package.

Examples

## Not run: 
library(rmarkdown)
draft("MyArticle.Rmd", template = "ctex", package = "rticles")

## End(Not run)

[Package rticles version 0.6 Index]