List available journals

List available journal names in this package.

journals()

Value

A character vector of the journal names.

Details

These names can be useful in two ways:

You can add _article suffix to get the name of the output format (e.g., rjournal_article()).

You can use the name directly in the template argument of rmarkdown::draft().

Examples

rticles::journals()
#>  [1] "acm"            "acs"            "aea"            "agu"           
#>  [5] "amq"            "ams"            "arxiv"          "asa"           
#>  [9] "bioinformatics" "biometrics"     "copernicus"     "ctex"          
#> [13] "elsevier"       "frontiers"      "ieee"           "ims"           
#> [17] "jasa"           "joss"           "jss"            "lipics"        
#> [21] "mdpi"           "mnras"          "oup"            "peerj"         
#> [25] "pihph"          "plos"           "pnas"           "rjournal"      
#> [29] "rsos"           "rss"            "sage"           "sim"           
#> [33] "springer"       "tf"