jss_article {rticles}	R Documentation

Journal of Statistical Software (JSS) format.

Description

Format for creating a Journal of Statistical Software (JSS) articles. Adapted from http://www.jstatsoft.org/about/submissions.

Usage

jss_article(..., keep_tex = TRUE, citation_package = "natbib")

Arguments

...
Arguments to rmarkdown::pdf_document

keep_tex
Keep the intermediate tex file used in the conversion to PDF

citation_package
The LaTeX package to process citations, natbib or biblatex. Use none if neither package is to be used.

[Package rticles version 0.9 Index]