mnras_article {rticles}	R Documentation

Monthly Notices of the Royal Astronomical Society (MNRAS) Journal format.

Description

Format for creating an Monthly Notices of Royal Astronomical Society (MNRAS) Journal articles. Adapted from https://www.ras.org.uk/news-and-press/2641-new-version-of-the-mnras-latex-package.

Usage

mnras_article(..., keep_tex = TRUE, md_extensions = c(),
  fig_caption = TRUE)

Arguments

...
Arguments to rmarkdown::pdf_document

keep_tex
Keep the intermediate tex file used in the conversion to PDF

md_extensions
Markdown extensions to be added or removed from the default definition or R Markdown. See the rmarkdown_format for additional details.

fig_caption
TRUE to render figures with captions

Value

R Markdown output format to pass to render

Examples

## Not run: 
library(rmarkdown)
draft("MyArticle.Rmd", template = "mnras_article", package = "rticles")

## End(Not run)

[Package rticles version 0.6 Index]