peerj_article {rticles}	R Documentation

PeerJ journal format.

Description

Format for creating submissions to The PeerJ.

Usage

peerj_article(..., keep_tex = TRUE, citation_package = "none",
  base_format = rmarkdown::pdf_document)

Arguments

...
Additional arguments to base_format

keep_tex
Keep the intermediate tex file used in the conversion to PDF

citation_package
The LaTeX package to process citations, natbib or biblatex. Use none if neither package is to be used.

base_format
The function to use for the base format of the article. By default, this is rmarkdown::pdf_document, but to use bookdown's cross-referencing feature, this can be set to bookdown::pdf_document2

Details

This was adapted from the PeerJ Overleaf Template.

Value

R Markdown output format to pass to render

Note

If you use rmarkdown::pdf_document(), all internal references (i.e. tables and figures) must use \ref{} whereas with bookdown::pdf_document2(), you can additionally use \@ref().

Examples

## Not run: 
library(rmarkdown)
draft("MyArticle.Rmd", template = "peerj_article", package = "rticles")

## End(Not run)

[Package rticles version 0.6 Index]