rss_article {rticles}	R Documentation

Royal Statistical Society Journal Format

Description

Format for creating articles for Royal Statistical Society Adapted from https://www.rss.org.uk/RSS/Publications/Journals/Journals_get_involved/RSS/Publications/Journals_sub/Get_Involved.aspx

Usage

rss_article(..., keep_tex = TRUE, citation_package = "natbib")

Arguments

...
Arguments to rmarkdown::pdf_document

keep_tex
Keep the intermediate tex file used in the conversion to PDF

citation_package
The LaTeX package to process citations, natbib or biblatex. Use none if neither package is to be used.

Value

R Markdown output format to pass to render

[Package rticles version 0.6 Index]