sage_article {rticles}	R Documentation

Sage Journals format.

Description

Format for creating submissions to Sage Journals. Based on the official Sage Journals class.

Usage

sage_article(..., highlight = NULL, citation_package = "natbib")

Arguments

...
Additional arguments to rmarkdown::pdf_document

highlight
Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock". Pass NULL to prevent syntax highlighting.

citation_package
The LaTeX package to process citations, natbib or biblatex. Use none if neither package is to be used.

Details

Possible arguments for the YAML header are:

title title of the manuscript

runninghead short author list for header

author list of authors, containing name and num

address list containing num and org for defining author affiliations

corrauth corresponding author name and address

email correspondence email

abstract abstract, limited to 200 words

keywords keywords for the artucle

bibliography BibTeX .bib file name

classoption options of the sagej class

header-includes: custom additions to the header, before the \begin{document} statement

include-after: for including additional LaTeX code before the \end{document} statement

Value

R Markdown output format to pass to render

Examples

## Not run: 
library(rmarkdown)
draft("MyArticle.Rmd", template = "sage_article", package = "rticles")

## End(Not run)

[Package rticles version 0.6 Index]