sim_article {rticles}	R Documentation

Statistics in Medicine format.

Description

Format for creating submissions to Statistics in Medicine. Based on the official Statistics in Medicine class.

Usage

sim_article(..., highlight = NULL, citation_package = "natbib")

Arguments

...
Additional arguments to rmarkdown::pdf_document

highlight
Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock". Pass NULL to prevent syntax highlighting.

citation_package
The LaTeX package to process citations, natbib or biblatex. Use none if neither package is to be used.

Details

Possible arguments for the YAML header are:

title title of the manuscript

author list of authors, containing name and num

address list containing num and org for defining author affiliations

presentaddress not sure what they mean with this

corres author and address for correspondence

authormark short author list for header

received, revised, accepted dates of submission, revision, and acceptance of the manuscript

abstract abstract, limited to 250 words

keywords up to 6 keywords

bibliography BibTeX .bib file

classoption options of the WileyNJD-v2 class

longtable set to true to include the longtable package, used by default from pandoc to convert markdown to LaTeX code

header-includes: custom additions to the header, before the \begin{document} statement

include-after: for including additional LaTeX code before the \end{document} statement

Value

R Markdown output format to pass to render

Examples

## Not run: 
library(rmarkdown)
draft("MyArticle.Rmd", template = "sim_article", package = "rticles")

## End(Not run)

[Package rticles version 0.6 Index]