R Markdown output formats for (journal) articles

Most article formats are based on rmarkdown::pdf_document(), with a custom Pandoc LaTeX template and different default values for other arguments (e.g., keep_tex = TRUE).

acm_article(...)

acs_article(
  ...,
  keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris"),
  fig_caption = TRUE
)

aea_article(..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris"))

agu_article(
  ...,
  keep_tex = TRUE,
  citation_package = "natbib",
  highlight = NULL,
  md_extensions = c("-autolink_bare_uris", "-auto_identifiers")
)

amq_article(
  ...,
  latex_engine = "xelatex",
  keep_tex = TRUE,
  fig_caption = TRUE,
  md_extensions = c("-autolink_bare_uris")
)

ams_article(..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris"))

asa_article(..., keep_tex = TRUE, citation_package = "natbib")

arxiv_article(..., keep_tex = TRUE)

bioinformatics_article(..., keep_tex = TRUE, citation_package = "natbib")

biometrics_article(..., keep_tex = TRUE, citation_package = "natbib")

ctex_article(..., template = "default", latex_engine = "xelatex")

ctex(..., template = "default", latex_engine = "xelatex")

elsevier_article(
  ...,
  keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris")
)

frontiers_article(..., keep_tex = TRUE)

ims_article(
  journal = c("aoas", "aap", "aop", "aos", "sts"),
  keep_tex = TRUE,
  citation_package = "natbib",
  md_extensions = c("-autolink_bare_uris"),
  pandoc_args = NULL,
  ...
)

jasa_article(
  ...,
  keep_tex = TRUE,
  latex_engine = "xelatex",
  citation_package = "natbib"
)

lipics_article(
  ...,
  latex_engine = "xelatex",
  keep_tex = TRUE,
  citation_package = "natbib",
  md_extensions = c("-autolink_bare_uris", "-auto_identifiers")
)

mdpi_article(..., keep_tex = TRUE)

mnras_article(..., keep_tex = TRUE, fig_caption = TRUE)

oup_article(..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris"))

peerj_article(..., keep_tex = TRUE)

pihph_article(
  ...,
  keep_tex = TRUE,
  latex_engine = "xelatex",
  citation_package = "biblatex"
)

plos_article(..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris"))

pnas_article(..., keep_tex = TRUE)

sage_article(..., highlight = NULL, citation_package = "natbib")

sim_article(..., highlight = NULL, citation_package = "natbib")

springer_article(..., keep_tex = TRUE, citation_package = "default")

tf_article(..., keep_tex = TRUE, citation_package = "natbib")

Arguments

..., keep_tex, latex_engine, citation_package, highlight, fig_caption, md_extensions, template, pandoc_args
Arguments passed to rmarkdown::pdf_document().

journal
one of "aoas", "aap", "aop", "aos", "sts" for ims_article

Value

An R Markdown output format.

Details

You can find more details about each output format below.

acm_article

Format for creating an Association for Computing Machinery (ACM) articles. Adapted from https://www.acm.org/publications/proceedings-template.

acs_article

Format for creating an American Chemical Society (ACS) Journal articles. Adapted from https://pubs.acs.org/page/4authors/submission/tex.html.

aea_article

Format for creating submissions to the American Economic Association (AER, AEJ, JEL, PP).

agu_article

Format for creating a American Geophysical Union (AGU) article. Adapted from https://www.agu.org/Publish-with-AGU/Publish/#1.

amq_article

Ce format a été adapté du format du bulletin de l'AMQ.

ams_article

Format for creating an American Meteorological Society (AMS) Journal articles. Adapted from https://www.ametsoc.org/ams/index.cfm/publications/authors/journal-and-bams-authors/author-resources/latex-author-info/.

asa_article

This format was adapted from The American Statistican (TAS) format, but it should be fairly consistent across American Statistical Association (ASA) journals.

arxiv_article

Adapted from the George Kour's format for arXiv and bio-arXiv preprints. So far as I'm aware, entirely unofficial but still a staple.

bioinformatics_article

Format for creating submissions to a Bioinformatics journal. Adapted from https://academic.oup.com/bioinformatics/pages/submission_online.

biometrics_article

This format was adapted from the Biometrics journal.

ctex_article

A wrapper function for rmarkdown::pdf_document() and the default value of latex_engine is changed to xelatex, so it works better for typesetting Chinese documents with the LaTeX package ctex. The function ctex is an alias of ctex_article.

elsevier_article

Format for creating submissions to Elsevier journals. Adapted from https://www.elsevier.com/authors/policies-and-guidelines/latex-instructions.

frontiers_article

Format for creating Frontiers journal articles. Adapted from https://www.frontiersin.org/about/author-guidelines.

ims_article

Format for creating submissions to the Institute of Mathematical Statistics IMS journals and publications. Adapted from https://github.com/vtex-soft/texsupport.ims-aoas.

The argument journal accepts the acronym of any of the journals in IMS:

aap: The Annals of Applied Probability

aoas: The Annals of Applied Statistics

aop: The Annals of Probability

aos: The Annals of Statistics

sts: Statistical Science

jasa_article

Format for creating submissions to the Journal of the Acoustical Society of America. Adapted from https://acousticalsociety.org/preparing-latex-manuscripts/.

lipics_article

Format for creating submissions to LIPIcs - Leibniz International Proceedings Informatics - articles. Adapted from the official Instructions for Authors at https://submission.dagstuhl.de/documentation/authors and the template from the archive authors-lipics-v2019.zip downloaded with version tag v2019.2. The template is provided under The LaTeX Project Public License (LPPL), Version 1.3c.

mdpi_article

Format for creating submissions to Multidisciplinary Digital Publishing Institute (MDPI) journals. Adapted from https://www.mdpi.com/authors/latex.

mnras_article

Format for creating an Monthly Notices of Royal Astronomical Society (MNRAS) Journal articles. Adapted from https://ras.ac.uk.

oup_article

Format for creating submissions to many Oxford University Press journals. Adapted from https://academic.oup.com/journals/pages/authors/preparing_your_manuscript and https://academic.oup.com/icesjms/pages/General_Instructions.

peerj_article

Format for creating submissions to The PeerJ Journal. This was adapted from the PeerJ Overleaf Template.

pihph_article

Format for creating submissions to the Papers in Historical Phonology (http://journals.ed.ac.uk/pihph/about/submissions). Adapted from https://github.com/pihph/templates. This format works well with latex_engine = "xelatex" and citation_package="biblatex", which are the default. It may not work correctly if you change these value. In that case, please open an issue and, a PR to contribute a change in the template.

plos_article

Format for creating submissions to PLOS journals. Adapted from https://journals.plos.org/ploscompbiol/s/latex.

pnas_article

Format for creating submissions to PNAS journals.

sage_article

Format for creating submissions to Sage Journals. Based on the official Sage Journals https://uk.sagepub.com/sites/default/files/sage_latex_template_4.zipclass.

Possible arguments for the YAML header are:

title title of the manuscript

runninghead short author list for header

author list of authors, containing name and num

address list containing num and org for defining author affiliations

corrauth corresponding author name and address

email correspondence email

abstract abstract, limited to 200 words

keywords keywords for the article

bibliography BibTeX .bib file name

classoption options of the sagej class

header-includes: custom additions to the header, before the \begin{document} statement

include-after: for including additional LaTeX code before the \end{document} statement

sim_article

Format for creating submissions to Statistics in Medicine. Based on the official Statistics in Medicine class.

Possible arguments for the YAML header are:

title title of the manuscript

author list of authors, containing name and num

address list containing num and org for defining author affiliations

presentaddress not sure what they mean with this

corres author and address for correspondence

authormark short author list for header

received, revised, accepted dates of submission, revision, and acceptance of the manuscript

abstract abstract, limited to 250 words

keywords up to 6 keywords

bibliography BibTeX .bib file

classoption options of the WileyNJD-v2 class

longtable set to true to include the longtable package, used by default from pandoc to convert markdown to LaTeX code

header-includes: custom additions to the header, before the \begin{document} statement

include-after: for including additional LaTeX code before the \end{document} statement

springer_article

This format was adapted from the Springer Macro package for Springer Journals.

tf_article

Format for creating submissions to a Taylor & Francis journal. Adapted from https://www.tandf.co.uk/journals/authors/InteractCADLaTeX.zip.

Examples

if (FALSE) {
rmarkdown::draft("MyArticle.Rmd", template = "acm", package = "rticles")
rmarkdown::draft("MyArticle.Rmd", template = "asa", package = "rticles")}