acm_article {rticles}	R Documentation

R Markdown output formats for (journal) articles

Description

Most article formats are based on rmarkdown::pdf_document(), with a custom Pandoc LaTeX template and different default values for other arguments (e.g., keep_tex = TRUE).

Usage

acm_article(...)

acs_article(..., keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris"), fig_caption = TRUE)

aea_article(..., keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris"))

agu_article(..., keep_tex = TRUE, citation_package = "natbib",
  highlight = NULL, md_extensions = c("-autolink_bare_uris",
  "-auto_identifiers"))

amq_article(..., latex_engine = "xelatex", keep_tex = TRUE,
  fig_caption = TRUE, md_extensions = c("-autolink_bare_uris"))

ams_article(..., keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris"))

asa_article(..., keep_tex = TRUE, citation_package = "natbib")

biometrics_article(..., keep_tex = TRUE, citation_package = "natbib")

ctex(..., latex_engine = "xelatex")

elsevier_article(..., keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris"))

frontiers_article(..., keep_tex = TRUE)

mdpi_article(..., keep_tex = TRUE)

mnras_article(..., keep_tex = TRUE, fig_caption = TRUE)

peerj_article(..., keep_tex = TRUE)

plos_article(..., keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris"))

pnas_article(..., keep_tex = TRUE)

sage_article(..., highlight = NULL, citation_package = "natbib")

sim_article(..., highlight = NULL, citation_package = "natbib")

springer_article(..., keep_tex = TRUE, citation_package = "none")

tf_article(..., keep_tex = TRUE, citation_package = "natbib")

Arguments

..., keep_tex, latex_engine, citation_package, highlight, fig_caption, md_extensions
Arguments passed to rmarkdown::pdf_document().

Value

An R Markdown output format.

Details

You can find more details about each output format below.

acm_article

Format for creating an Association for Computing Machinery (ACM) articles. Adapted from http://www.acm.org/publications/article-templates/proceedings-template.html.

acs_article

Format for creating an American Chemical Society (ACS) Journal articles. Adapted from http://pubs.acs.org/page/4authors/submission/tex.html.

aea_article

Format for creating submissions to the American Economic Association (AER, AEJ, JEL, PP).

agu_article

Format for creating a American Geophysical Union (AGU) article. Adapted from https://publications.agu.org/author-resource-center/checklists-and-templates/.

amq_article

Ce format a été adapté du format du bulletin de l'AMQ.

ams_article

Format for creating an American Meteorological Society (AMS) Journal articles. Adapted from https://www.ametsoc.org/ams/index.cfm/publications/authors/journal-and-bams-authors/author-resources/latex-author-info/.

asa_article

This format was adapted from The American Statistican (TAS) format, but it should be fairly consistent across American Statistical Association (ASA) journals.

biometrics_article

This format was adapted from the Biometrics journal.

ctex

A wrapper function for rmarkdown::pdf_document() and changed the default values of two arguments template and latex_engine so it works better for typesetting Chinese documents with the ctex LaTeX package.

elsevier_article

Format for creating submissions to Elsevier journals. Adapted from https://www.elsevier.com/authors/author-schemas/latex-instructions.

frontiers_article

Format for creating Frontiers journal articles. Adapted from http://home.frontiersin.org/about/author-guidelines.

mdpi_article

Format for creating submissions to Multidisciplinary Digital Publishing Institute (MDPI) journals. Adapted from http://www.mdpi.com/authors/latex.

mnras_article

Format for creating an Monthly Notices of Royal Astronomical Society (MNRAS) Journal articles. Adapted from https://www.ras.org.uk/news-and-press/2641-new-version-of-the-mnras-latex-package.

peerj_article

Format for creating submissions to The PeerJ Journal. This was adapted from the PeerJ Overleaf Template.

plos_article

Format for creating submissions to PLOS journals. Adapted from http://journals.plos.org/ploscompbiol/s/latex.

pnas_article

Format for creating submissions to PNAS journals.

sage_article

Format for creating submissions to Sage Journals. Based on the official Sage Journals class.

Possible arguments for the YAML header are:

title title of the manuscript

runninghead short author list for header

author list of authors, containing name and num

address list containing num and org for defining author affiliations

corrauth corresponding author name and address

email correspondence email

abstract abstract, limited to 200 words

keywords keywords for the article

bibliography BibTeX .bib file name

classoption options of the sagej class

header-includes: custom additions to the header, before the \begin{document} statement

include-after: for including additional LaTeX code before the \end{document} statement

sim_article

Format for creating submissions to Statistics in Medicine. Based on the official Statistics in Medicine class.

Possible arguments for the YAML header are:

title title of the manuscript

author list of authors, containing name and num

address list containing num and org for defining author affiliations

presentaddress not sure what they mean with this

corres author and address for correspondence

authormark short author list for header

received, revised, accepted dates of submission, revision, and acceptance of the manuscript

abstract abstract, limited to 250 words

keywords up to 6 keywords

bibliography BibTeX .bib file

classoption options of the WileyNJD-v2 class

longtable set to true to include the longtable package, used by default from pandoc to convert markdown to LaTeX code

header-includes: custom additions to the header, before the \begin{document} statement

include-after: for including additional LaTeX code before the \end{document} statement

springer_article

This format was adapted from the Springer Macro package for Springer Journals.

tf_article

Format for creating submissions to a Taylor & Francis journal. Adapted from http://www.tandf.co.uk/journals/authors/InteractCADLaTeX.zip.

Examples

## Not run: 
rmarkdown::draft("MyArticle.Rmd", template = "acm_article", package = "rticles")
rmarkdown::draft("MyArticle.Rmd", template = "asa_article", package = "rticles")
## End(Not run)

[Package rticles version 0.12 Index]