copernicus_article {rticles}	R Documentation

Copernicus journals format.

Description

Format for creating submissions to Copernicus journals.

Usage

copernicus_article(..., keep_tex = TRUE, citation_package = "natbib",
  md_extensions = c("-autolink_bare_uris", "-auto_identifiers"))

copernicus_journal_abbreviations(journal_name = "*")

Arguments

...
Additional arguments to rmarkdown::pdf_document().

keep_tex
Keep the intermediate tex file used in the conversion to PDF

citation_package
The LaTeX package to process citations, natbib or biblatex. Use none if neither package is to be used.

md_extensions
Markdown extensions to be added or removed from the default definition or R Markdown. See the rmarkdown_format for additional details.

journal_name
A regular expression to filter the by the journal name, see pattern in grep; defaults to *.

Details

This was adapted from https://publications.copernicus.org/for_authors/manuscript_preparation.html.

An number of required and optional manuscript sections, e.g. acknowledgements, competinginterests, or authorcontribution, must be declared using the respective properties of the R Markdown header - see skeleton file.

Version: Based on copernicus_package.zip in the version 5.3, 18 February 2019, using copernicus.cls in version 8.82.

Copernicus journal abbreviations: You can use the function copernicus_journal_abbreviations() to get the journal abbreviation for all journals supported by the copernicus article template.

Important note: The online guidelines by Copernicus are the official resource. Copernicus is not responsible for the community contributions made to support the template in this package. Copenicus converts all typeset TeX files into XML, the expressions and markups have to be highly standardized. Therefore, please keep the following in mind:

Please provide only one figure file for figures with several panels, and please do not use \subfloat or similar commands.

Please use only commands in which words, numbers, etc. are within braces (e.g. \textrm{TEXT} instead of {\rm TEXT}).

For algorithms, please use the syntax given in template.tex or provide your algorithm as a figure.

Please do not define new commands.

The most commonly used packages (\usepackage{}) are integrated in the copernicus.cls. Some other packages often used by the community are defined in template.tex. Please do not insert additional ones in your *.tex file.

Spaces in labels (\label{}) are not allowed; please make sure that no label name is assigned more than once.

Please do not use \paragraph{}; only \subsubsection{} is allowed.

It is not possible to add tables in colour.

Value

An R Markdown output format.

Note

If you use rmarkdown::pdf_document(), all internal references (i.e. tables and figures) must use \ref{} whereas with bookdown::pdf_document2(), you can additionally use \@ref().

References

Manuscript preparation guidelines for authors. https://publications.copernicus.org/for_authors/manuscript_preparation.html

Examples

names(copernicus_journal_abbreviations())
copernicus_journal_abbreviations(journal_name = "Science Data")
## Not run: 
library("rmarkdown")
draft("MyArticle.Rmd", template = "copernicus_article", package = "rticles")
render("MyArticle/MyArticle.Rmd")

## End(Not run)

[Package rticles version 0.9 Index]