ieee_article {rticles}	R Documentation

IEEE Transactions journal format.

Description

Format for creating submissions to IEEE Transaction journals. Adapted from http://www.ieee.org/publications_standards/publications/authors/author_templates.html.

Usage

ieee_article(
  draftmode = c("final", "draft", "draftcls", "draftclsnofoot"),
  hyphenfixes = "op-tical net-works semi-conduc-tor",
  IEEEspecialpaper = "",
  with_ifpdf = FALSE,
  with_cite = FALSE,
  with_amsmath = FALSE,
  with_algorithmic = FALSE,
  with_subfig = FALSE,
  with_array = FALSE,
  with_dblfloatfix = FALSE,
  keep_tex = TRUE,
  pandoc_args = NULL,
  md_extensions = c("-autolink_bare_uris"),
  ...
)

Arguments

draftmode
Specify the draft mode to control spacing and whether images should be rendered. Valid options are: "final" (default), "draft", "draftcls", or "draftclsnofoot".

hyphenfixes
A character value that provides the correct hyphenations for ambiguous words. Separate new words with spaces.

IEEEspecialpaper
A character value containing the publication's special paper designation.

with_ifpdf
A logical value turning on (TRUE) or off (FALSE) the ifpdf LaTeX package.

with_cite
A logical value turning on (TRUE) or off (FALSE) the cite LaTeX package.

with_amsmath
A logical value turning on (TRUE) or off (FALSE) the amsmath LaTeX package.

with_algorithmic
A logical value turning on (TRUE) or off (FALSE) the algorithmic LaTeX package.

with_subfig
A logical value turning on (TRUE) or off (FALSE) the subfig LaTeX package.

with_array
A logical value turning on (TRUE) or off (FALSE) the array LaTeX package.

with_dblfloatfix
A logical value turning on (TRUE) or off (FALSE) the dblfloatfix LaTeX package.

keep_tex
Keep the intermediate tex file used in the conversion to PDF

pandoc_args
Additional command line options to pass to pandoc

md_extensions
Markdown extensions to be added or removed from the default definition or R Markdown. See the rmarkdown_format for additional details.

...
Additional arguments to rmarkdown::pdf_document()

Details

Presently, only the "conference" paper mode offered by the IEEEtran.cls is supported.

References

Shell, Michael. "How to use the IEEEtran LATEX class." Journal of LATEX Class Files 1.11 (2002): 10-20. http://mirrors.rit.edu/CTAN/macros/latex/contrib/IEEEtran/IEEEtran_HOWTO.pdf

[Package rticles version 0.21 Index]