joss_article {rticles}	R Documentation

Journal of Open Source Software (JOSS) format.

Description

Format for creating a Journal of Open Source Software (JOSS) or Journal of Open Source Education (JOSE) articles. Adapted from https://github.com/openjournals/whedon. As these journals take articles as markdown, this format can be used to generate markdown from R Markdown and to locally preview how the article will appear as PDF.

Usage

joss_article(journal = "JOSS", keep_md = TRUE, latex_engine = "xelatex", ...)

Arguments

journal
one of "JOSS" or"JOSE"

keep_md
Whether to retain the intermediate markdown and images. Defaults to TRUE.

latex_engine, ...
Arguments passed to rmarkdown::pdf_document()

Details

The following variables may be set in YAML metadata to populate fields in the article PDF, but are only necccessary fo local preview: formatted_doi, citation_author, year, volume, issue, page, submitted, published, review_url, repository, and archive_doi.

[Package rticles version 0.21 Index]