journals {rticles}	R Documentation

List available journals

Description

List available journal names in this package.

Usage

journals()

Details

These names can be useful in two ways:

You can add _article suffix to get the name of the output format (e.g., rjournal_article()).

You can use the name directly in the template argument of rmarkdown::draft().

Value

A character vector of the journal names.

Examples

rticles::journals()

[Package rticles version 0.21 Index]