rsos_article {rticles}	R Documentation

Royal Society Open Science journal format.

Description

Format for creating submissions to Royal Society Open Science journals.

Usage

rsos_article(..., keep_tex = TRUE, latex_engine = "xelatex",
  pandoc_args = NULL, includes = NULL, fig_crop = TRUE)

Arguments

...
Additional arguments to rmarkdown::pdf_document

keep_tex
Keep the intermediate tex file used in the conversion to PDF

latex_engine
LaTeX engine for producing PDF output. Options are "pdflatex", "lualatex", and "xelatex".

pandoc_args
Additional command line options to pass to pandoc

includes
Named list of additional content to include within the document (typically created using the includes function).

fig_crop
TRUE to automatically apply the pdfcrop utility (if available) to pdf figures

Author(s)

Thierry Onkelinx, thierry.onkelinx@inbo.be

[Package rticles version 0.12 Index]